% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cusum_prop_r_t_EM_smoothing_spline}
\alias{cusum_prop_r_t_EM_smoothing_spline}
\title{Function to calculate the CUSUM chart at different time points given by t_grid}
\usage{
cusum_prop_r_t_EM_smoothing_spline(
  cumulative_baseline_haz_vec_general_func,
  baseline_haz_vec_general_func,
  start_year,
  age_vec,
  gender_vec,
  x_matrix,
  time_obs_vec,
  arrival_time_vec,
  delta_i_vec,
  beta_vec,
  rho,
  t_grid,
  pop_data_male,
  pop_data_female,
  end_year_table
)
}
\arguments{
\item{cumulative_baseline_haz_vec_general_func}{R function that corresponds to the estimated cumulative baseline hazard obtained by the excess hazard model based on EM algorithm.}

\item{baseline_haz_vec_general_func}{R function that corresponds to the estimated baseline hazard obtained by the excess hazard model based on EM algorithm.}

\item{start_year}{Calendar year of when the monitoring starts. End year is implicitly defined via t_grid.}

\item{age_vec}{A vector containing age of the observations when they arrive in the study.}

\item{gender_vec}{A vector containing gender indicator - 1 for male, 2 for female.}

\item{x_matrix}{A matrix containing the predictor variables (need to contain the same amount of columns as the dimension of beta_vec, the estimated parameter vector).}

\item{time_obs_vec}{A vector containing the observed time of the observations.}

\item{arrival_time_vec}{A vector containing the amount of time units after start year the observations arrive. Should be less than or equal to max(t_grid).}

\item{delta_i_vec}{A vector containing the event indicator at the end of follow-up.}

\item{beta_vec}{A vector containing the estimated parameters from the model related to the predictor variables.}

\item{rho}{Proportional constant for the out-of-control hazard model.}

\item{t_grid}{The time grid to calculate the CUSUM chart.}

\item{pop_data_male}{A matrix of population table for male. Should follow the matrix format with first column being the year, second column being age and third column being the hazard values.}

\item{pop_data_female}{Similar as above, but for female.}

\item{end_year_table}{The last calendar year appearing in the population table.}
}
\value{
final_result_matrix A matrix where the first column is the chart value, the second column counts the number of events.
}
\description{
Function to calculate the CUSUM chart at different time points given by t_grid using proportional alternative with output from rsadd(..., method = "EM") as baseline hazard
}
