\name{CUSUMrelsurv-package}
\alias{CUSUMrelsurv-package}
\alias{CUSUMrelsurv}
\docType{package}
\title{
\packageTitle{CUSUMrelsurv}
}
\description{
\packageDescription{CUSUMrelsurv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CUSUMrelsurv}
\packageIndices{CUSUMrelsurv}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CUSUMrelsurv}

Maintainer: \packageMaintainer{CUSUMrelsurv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
