% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sim_bootstrap_cusum_additive_excess_non_negative_threshold_piecewise_function_parallel.R
\name{sim_bootstrap_cusum_add_excess_non_neg_threshold_piecewise_function_parallel}
\alias{sim_bootstrap_cusum_add_excess_non_neg_threshold_piecewise_function_parallel}
\title{Function to simulate threshold for CUSUM with additive alternative assuming non-negative excess hazard under piecewise constant baseline hazard.}
\usage{
sim_bootstrap_cusum_add_excess_non_neg_threshold_piecewise_function_parallel(
  partition_t_vec,
  baseline_vec,
  n_iterations,
  n_cores = NULL,
  start_year,
  estimated_rate_censoring,
  age_vec,
  gender_vec,
  x_matrix,
  arrival_time_vec,
  beta_vec,
  gamma_vec,
  t_grid,
  pop_data_male,
  pop_data_female,
  end_year_table,
  random_state = NULL
)
}
\arguments{
\item{partition_t_vec}{A vector that defines the partition of the time period for the piecewise constant baseline hazard.}

\item{baseline_vec}{A vector of baseline hazard values defined over the time partition.}

\item{n_iterations}{Number of simulated CUSUM charts under in-control scenario for calculating threshold.}

\item{n_cores}{Number of cores to run in parallel.}

\item{start_year}{Calendar year of when the monitoring starts. End year is implicitly defined via t_grid.}

\item{estimated_rate_censoring}{Estimated censoring rate when simulating interim censoring times for observations.}

\item{age_vec}{A vector containing age of the observations when they arrive in the study.}

\item{gender_vec}{A vector containing gender indicator - 0 for male, 1 for female.}

\item{x_matrix}{A matrix containing the predictor variables (need to contain the same amount of columns as the dimension of beta_vec, the estimated parameter vector).}

\item{arrival_time_vec}{A vector containing the amount of time units after start year the observations arrive. Should be less than or equal to max(t_grid).}

\item{beta_vec}{A vector containing the estimated parameters from the model related to the predictor variables.}

\item{gamma_vec}{A vector of values of the additive constant for the out-of-control hazard model.}

\item{t_grid}{The time grid to calculate the CUSUM chart.}

\item{pop_data_male}{A matrix of population table for male. Should follow the matrix format with first column being the year, second column being age and third column being the hazard values.}

\item{pop_data_female}{Similar as above, but for female.}

\item{end_year_table}{The last calendar year appearing in the population table.}

\item{random_state}{Seed value for reproducibility.}

\item{time_obs_vec}{A vector containing the observed time of the observations.}
}
\value{
store_matrix A matrix with the number of columns equal to the number of rho values. The number of rows is 'n_iterations'. The matrix stores the max value of all the simulated charts, which can be used to estimate the threshold.
}
\description{
Function to simulate threshold for CUSUM with additive alternative assuming non-negative excess hazard under piecewise constant baseline hazard (with covariates simulated via bootstrapping)
}
