% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportional_sim_predefined_obs_func_cusum_threshold_piecewise_function_parallel.R
\name{sim_predefined_obs_func_cusum_prop_threshold_piecewise_function_parallel}
\alias{sim_predefined_obs_func_cusum_prop_threshold_piecewise_function_parallel}
\title{Function to simulate threshold for CUSUM with proportional alternative under piecewise constant baseline hazard.}
\usage{
sim_predefined_obs_func_cusum_prop_threshold_piecewise_function_parallel(
  sim_obs_func,
  base_piecewise_model = NULL,
  n_iterations,
  n_cores = NULL,
  rho_vec,
  pop_data_male,
  pop_data_female,
  end_year_table,
  random_state = NULL
)
}
\arguments{
\item{sim_obs_func}{A function that simulates covariates etc. under prespecified (true) covariate distribution.}

\item{base_piecewise_model}{Fitted model object from relsurv using piecewise constant baseline hazard.}

\item{n_iterations}{Number of simulated CUSUM charts under in-control scenario for calculating threshold.}

\item{n_cores}{Number of cores to run in parallel.}

\item{rho_vec}{A vector of values of the proportional constant for the out-of-control hazard model.}

\item{pop_data_male}{A matrix of population table for male. Should follow the matrix format with first column being the year, second column being age and third column being the hazard values.}

\item{pop_data_female}{Similar as above, but for female.}

\item{end_year_table}{The last calendar year appearing in the population table.}

\item{random_state}{Seed value for reproducibility.}
}
\value{
store_matrix A matrix with the number of columns equal to the number of rho values. The number of rows is 'n_iterations'. The matrix stores the max value of all the simulated charts, which can be used to estimate the threshold.
}
\description{
Function to simulate threshold for CUSUM with proportional alternative under piecewise constant baseline hazard (with covariates simulating via prespecified instructions).
}
