% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_excess_sim_add_excess_non_neg_piecewise_out}
\alias{vec_excess_sim_add_excess_non_neg_piecewise_out}
\title{Function to simulate excess times during a time period where the hazard starts in out-of-control}
\usage{
vec_excess_sim_add_excess_non_neg_piecewise_out(
  partition_t_vec,
  baseline_vec,
  max_follow_up_vec,
  u_vec,
  linear_predictor_vec,
  gamma
)
}
\arguments{
\item{partition_t_vec}{A vector that defines the partition of the time period for the piecewise constant baseline hazard.}

\item{baseline_vec}{A vector of baseline hazard values defined over the time partition.}

\item{max_follow_up_vec}{A vector corresponding to the max follow up times of the observations.}

\item{u_vec}{A vector of samples from Unif(0, 1) for inverse transform algorithm.}

\item{linear_predictor_vec}{the matrix vector product of the covariate matrix and the regression coefficients.}

\item{gamma}{Additive constant for the out-of-control hazard model.}
}
\value{
time_excess_sim A vector of simulated excess times based on the input of the observations.
}
\description{
Function to simulate excess times during a time period where the hazard starts in out-of-control using additive alternative, assuming non-negative excess hazard
}
