% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_excess_sim_EM}
\alias{vec_excess_sim_EM}
\title{Function to simulate excess times based on output from semi-parametric (EM) excess hazard model}
\usage{
vec_excess_sim_EM(
  predict_cumulative_baseline_haz_general_func,
  min_cumulative_baseline_event_time,
  u_vec,
  linear_predictor_vec,
  max_follow_up_vec
)
}
\arguments{
\item{predict_cumulative_baseline_haz_general_func}{A function that returns predicted value of cumulative baseline based on the smoothed estimated from the EM output from relsurv.}

\item{min_cumulative_baseline_event_time}{The smallest time to event that was used in the semi-parametric model for estimation.}

\item{u_vec}{A vector of samples from Unif(0, 1) for inverse transform algorithm.}

\item{linear_predictor_vec}{the matrix vector product of the covariate matrix and the regression coefficients.}

\item{max_follow_up_vec}{A vector corresponding to the max follow up times of the observations.}
}
\value{
time_excess_sim A vector of simulated excess times based on the input of the observations.
}
\description{
Function to simulate excess times based on output from semi-parametric (EM) excess hazard model
}
