% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_pop_sim}
\alias{vec_pop_sim}
\title{Function to simulate population times based on life tables}
\usage{
vec_pop_sim(
  age_vec,
  gender_vec,
  start_year,
  end_year,
  table_end_year,
  arrival_time_vec,
  pop_matrix_male,
  pop_matrix_female,
  u
)
}
\arguments{
\item{age_vec}{A vector containing age of the observations when they arrive in the study.}

\item{gender_vec}{A vector containing gender indicator - 0 for male, 1 for female.}

\item{start_year}{Calendar year of when the monitoring starts.}

\item{end_year}{Calendar year of when the monitoring ends.}

\item{table_end_year}{The last calendar year appearing in the population table.}

\item{arrival_time_vec}{A vector containing the amount of time units after start year the observations arrive.}

\item{pop_matrix_male}{Population table for male, should be a matrix with first column corresponding to year, second column to age and third column to the population hazard value.}

\item{pop_matrix_female}{Population table for female, should be a matrix with first column corresponding to year, second column to age and third column to the population hazard value.}

\item{u}{A vector of samples from Unif(0, 1) for inverse transform algorithm.}
}
\value{
A matrix where first column corresponds to the population time and the second is the event indicator.
}
\description{
Function to simulate population times based on life tables
}
